refactor(cockpit): extract review map renderer
